Remove .gitignore files
Since extensions are not used as composer.json roots anymore there is no
need for .gitignore to be specified in every extension. The same applies
to the nodejs packages.

The php-libraries/composer-project/devportal/.gitignore is special. This
is used as the initial root .gitignore file on newly created projects.

The Monorepo's root .gitignore file was tidied up, and the git ignore
rules that were still relevent were moved into the root .gitignore file.

Change-Id: Ia5daf554fcf198d81251177b2767a50845582cc7
Issue-Link: https://project.pronovix.net/issues/15749